changes in campaign flow added multi currrency support, keeping CYN as base currency added campaing settings in appsettings

+++ RenewalWebsite/Data/Migrations/20170811092112_DonationAmount.cs
@@ -29,7 +29,7 @@ namespace RenewalWebsite.Data.Migrations
                 name: "DonationAmount",
                 table: "Donations",
                 nullable: true,
-                oldClrType: typeof(int),
+                oldClrType: typeof(decimal),
                 oldNullable: true);
 
             migrationBuilder.CreateTable(
